Reorganize "reparsing" infrastructure
We will need to re-parse not just macros, but in some other cases also (for
example, when recognizing that a struct field is actually declared using a
macro). This paves the way for doing that.
