+++ hs-bindgen/src/HsBindgen/C/AST.hs
@@ -36,7 +36,7 @@ module HsBindgen.C.AST (
     -- ** Classification
   , isIncludeGuard
     -- ** Unrecognized
-  , UnrecognizedMacro(..)
+  , ReparseError(..)
   , Token(..)
   , TokenSpelling(..)
     -- * Source locations
@@ -53,9 +53,9 @@ import HsBindgen.C.AST.Literal
 import HsBindgen.C.AST.Macro
 import HsBindgen.C.AST.Name
 import HsBindgen.C.AST.Type
-import HsBindgen.C.Parser.Macro (UnrecognizedMacro(..))
-import HsBindgen.Clang.Util.Tokens
+import HsBindgen.C.Reparse.Infra (ReparseError(..))
 import HsBindgen.Clang.Util.SourceLoc.Type
+import HsBindgen.Clang.Util.Tokens
 
 {-------------------------------------------------------------------------------
   Top-level
@@ -73,7 +73,7 @@ data Decl =
     DeclStruct Struct
   | DeclTypedef Typedef
   | DeclEnum Enu
-  | DeclMacro (Either UnrecognizedMacro Macro)
+  | DeclMacro (Either ReparseError Macro)
   deriving stock (Show, Eq, Generic)
   deriving anyclass (PrettyVal)
