+++ hs-bindgen/tests/Orphans.hs
@@ -35,13 +35,13 @@ instance ToExpr C.MTerm
 instance ToExpr C.MultiLoc
 instance ToExpr C.PrimSign
 instance ToExpr C.PrimType
+instance ToExpr C.ReparseError
 instance ToExpr C.SingleLoc
 instance ToExpr C.Struct
 instance ToExpr C.StructField
 instance ToExpr C.TokenSpelling
 instance ToExpr C.Typ
 instance ToExpr C.Typedef
-instance ToExpr C.UnrecognizedMacro
 
 instance ToExpr a => ToExpr (C.Literal a)
 instance ToExpr a => ToExpr (C.Range a)
